% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv2indicatordata.R
\name{conv2indicatordata}
\alias{conv2indicatordata}
\title{Plotting indicator time series}
\usage{
conv2indicatordata(
  filename,
  labrows = 1,
  datecols = 1,
  indcols = 2,
  ulimcols = NA,
  llimcols = NA,
  default = F
)
}
\arguments{
\item{filename}{a .csv file with indicator names in the initial rows of the file, and dates and indicator values in columns below.
For default settings, see details below.}

\item{labrows}{a vector indicating the rows at the top of the file containing label information. Typically there are three rows with the indicator name, measurement unit, and sublabel.}

\item{datecols}{an integer which represents the column with the vector of dates.}

\item{indcols}{an integer of vector indicating which columns of the file contain the indicator values.}

\item{ulimcols}{an integer of vector indicating which columns of the file contain the upper limits of the confidence intervals, if display is desired.}

\item{llimcols}{an integer of vector indicating which columns of the file contain the lower limits of the confidence intervals, if display is desired.}

\item{default}{a logical value indicating whether the .csv file is in default format, in which case the other options do not need to

The "default" version of the .csv file should be formatted as follows:
\itemize{
\item \strong{Row 1} contains a character string of the indicator name (main title of plot).
\item \strong{Row 2} contains a character string of the indicator units (y-axis label).
\item \strong{Row 3} contains an optional character string with information on the sublabel (when multiple panels are used).
\item \strong{Column 1} is time values, \strong{columns 2} and on are indicator data.  The first three values of column 1 are NA.
\item Time can be in year (with century), or monthly time step in a variety of formats (e.g, Jan1986, Jan-86, 1986jan), including or excluding day of month.
}}
}
\description{
This function is for converting deprecated .csv files into an object of class 'indicatordata'.
}
