% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndicatorTimeSeries.r
\name{plotIndicatorTimeSeries}
\alias{plotIndicatorTimeSeries}
\title{Plotting indicator time series}
\usage{
plotIndicatorTimeSeries(
  filename,
  coltoplot = 2,
  plotrownum = 1,
  plotcolnum = 1,
  sublabel = F,
  sameYscale = F,
  yposadj = 1,
  widadj = 1,
  hgtadj = 1,
  type = "",
  trendAnalysis = T,
  propNAallow = 0.6,
  redgreen = T,
  anom = "none",
  outname = NA,
  outtype = "png"
)
}
\arguments{
\item{filename}{file in standardized indicator reporting csv format.
Column 1 is time values, columns 2+ are indicator data
Row 1 is indicator name (main title of plot)
Row 2 is units (y-axis label)
Row 3 is spatial extent or other specifying information
Time can be in yearly or monthly time step
- If annual time series, years must be in YYYY format (e.g., 2011)
- Monthly time series can be in any combination of text abbreviation
and year formats (e.g, Jan1986, Jan-86, 1986jan)
Note: must exclude days! (e.g, 1-Dec will be read as December 2001)}

\item{coltoplot}{column number of indicator file to plot (only col 2 is default)}

\item{plotrownum}{number of rows of plots in multi-panel plot}

\item{plotcolnum}{number of columns of plots in multi-panel plot
plotrownum and plotcolnum must be used if length(coltoplot) > 1
specify layout (e.g., 4 panels could be plotrownum = 2, plotcolnum = 2}

\item{sublabel}{whether extent description should appear within main label}

\item{sameYscale}{for multi-panel plots, if y-axis scale should be the same}

\item{yposadj}{manual adjustment of position of y-axis label}

\item{widadj}{adjust total width of plot}

\item{hgtadj}{adjust total height of plot}

\item{type}{plot type of lines or points
defaults to points, with lines for consecutive years only
for points only use type = "ptsOnly"
for all lines use type="allLines"}

\item{trendAnalysis}{whether to highlight trend in mean and SD over last 5 years
default is T unless fewer than 5 years of data}

\item{propNAallow}{proportion NAs allowed in trend analysis (defaults to 0.4)}

\item{redgreen}{whether or not to include red/green on plot (default is true)}

\item{anom}{calculate and plot monthly anomalies
for monthly anomalies: anom = "mon"
for standardized monthly anomalies: anom ="stmon"
@param outname specify alternate output filename (default is same as input)
@param outtype format for output (defaults to png, pdf also possible)}
}
\description{
This function is for standardized plotitng of indicator time series for Ecosystem Status Reports.
Typical formatting used by Alaska Report Card and CC IEA
Plots time series, values above and below 1 S.D., and mean
Highlights last 5 years of data and shows pattern in mean and trend
}
\examples{
 plotIndicatorTimeSeries("indicator.csv", coltoplot=2:4, plotrownum=3)
 plotIndicatorTimeSeries("amo.csv", coltoplot=2, plotrownum=1, plotcolnum=1, adjlabmain=1, sublabel=F)
 plotIndicatorTimeSeries("aveSocialConnectedness.csv", coltoplot=2, plotrownum=1, plotcolnum=1, sublabel=F)
 plotIndicatorTimeSeries("menhaden_abundance_index.csv", coltoplot=2, plotrownum=1, plotcolnum=1, sublabel=F, yposadj=0.8, widadj=0.8)
 plotIndicatorTimeSeries("seagrass_acreage.csv", coltoplot=2:7, plotrownum=3, plotcolnum=2, adjlabmain=1, sublabel=T, widadj=0.6, trendAnalysis=F)
function()
}
\keyword{}
