% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOrdScores.R
\name{plotOrdScores}
\alias{plotOrdScores}
\title{Plotting indicator time series}
\usage{
plotOrdScores(
  dataset,
  columns = NA,
  mintime = NA,
  maxtime = NA,
  noNAs = 0,
  color = "rainbow",
  lab.cols = 1,
  tim.cex = 1,
  add.pts = FALSE,
  indatformat = "\%Y",
  outdatformat = "\%y",
  main = "title",
  method = "pca",
  nmdsk = 2
)
}
\arguments{
\item{dataset}{A \link{data.frame} in which the first column is numerical values representing time steps
and subsequent columns are data values at those time steps.}

\item{columns}{an integer or vector defining the column numbers of data to be included.  Defaults to the entire data set.}

\item{mintime}{a value representing the earliest time step to be included in the analysis.  Defaults to first time step in data set.}

\item{maxtime}{a value representing the last time step to be included in the analysis.  Defaults to last time step in data set.}

\item{noNAs}{a value between 0 and 1 representing the proportion of allowable NAs for any given time series.
NA values are imputed with the mean across the time series for the ordination.  Time series exceeding the proportion of NAs
specified are removed from the analysis.}

\item{color}{color options for the segments connecting time steps. Either "rainbow" (default) or "gray" are available.}

\item{lab.cols}{color specification for the time step labels.}

\item{tim.cex}{a value indicating scaling for the size of the time step labels.}

\item{add.pts}{if TRUE, adds the PCA loadings or NMDS species on the ordination plot.  Defaults to FALSE.}

\item{indatformat}{a format in which the time steps are input, as defined in \link{strptime}.  Defaults to year with century ("\%Y").}

\item{outdatformat}{a format as defined in \link{strptime} which represents the desired format for labeling the plot.  Defaults to year without century ("\%y").}

\item{method}{the ordination method to be used in the analysis.  Defaults to principal components analysis ("pca") using \link{prcomp}.  Other option
is non-metric multidimensional scaling ("nmds") using \link{metaMDS}.}

\item{nmdsk}{an integer representing the number of dimensions to be used in \link{metaMDS}, if the NMDS method is specified.}
}
\description{
This function carries out ordination analysis and plots the ordination scores in two dimensions.
The input file is a data frame with a series of regular time steps and data values at those time steps.
}
\note{
Data must be input in a data frame where rows represent time steps and columns represent data series.
\itemize{
\item The first columns contains a string of time units in any specified format.
\item Columns 2+ contain data values for the respective time steps.
}
}
\examples{
## make traffic plot of Gulf of Mexico fishery landings
 
 plotOrdScores(GOMlandings)

}
\references{
Figure 1 in https://onlinelibrary.wiley.com/doi/full/10.1111/gcb.12894
}
