% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndicatorTimeSeries.r
\name{plotIndicatorTimeSeries}
\alias{plotIndicatorTimeSeries}
\title{Plotting indicator time series}
\usage{
plotIndicatorTimeSeries(
  filename,
  coltoplot = 2,
  plotrownum = 1,
  plotcolnum = 1,
  sublabel = F,
  sameYscale = F,
  yposadj = 1,
  widadj = 1,
  hgtadj = 1,
  type = "",
  trendAnalysis = T,
  propNAallow = 0.6,
  redgreen = T,
  anom = "none",
  outname = NA,
  outtype = "png"
)
}
\arguments{
\item{filename}{file in standardized indicator reporting csv format.
\strong{Column 1} is time values, \strong{columns 2} and on are indicator data.
\strong{Row 1} must be indicator name (main title of plot).
\strong{Row 2} must be units (y-axis label).
\strong{Row 3} is optional information for sublabel (when multiple panels are used).
Time can be in year (with century), or monthly time step in a variety of formats (e.g, Jan1986, Jan-86, 1986jan)}

\item{coltoplot}{column numbers of indicator file to plot (defaults to only column 2)}

\item{plotrownum}{number of rows of plots in multi-panel plot}

\item{plotcolnum}{number of columns of plots in multi-panel plot
plotrownum and plotcolnum must be used if length(coltoplot) > 1 to specify layout (e.g., 4 panels could be plotrownum = 2, plotcolnum = 2)}

\item{sublabel}{whether optional descriptive information should appear within main label}

\item{sameYscale}{set to TRUE for multi-panel plots if consistent y-axis scale is desired.}

\item{yposadj}{manual adjustment of position of y-axis label}

\item{widadj}{adjust total width of plot}

\item{hgtadj}{adjust total height of plot}

\item{type}{plot type of lines or points
defaults to points, with lines for consecutive years only
for points only use type = "ptsOnly"
for all lines use type="allLines"}

\item{trendAnalysis}{if TRUE, highlight trend in mean and SD over last 5 years; defaults to TRUE unless fewer than 5 years of data}

\item{propNAallow}{if proportion of missing values in last 5 years exceeds this value, overrides trendAnalysis=T and will not plot trend (defaults to 0.4)         -}

\item{redgreen}{set to FALSE to remove red/green shading of anomalies from plot}

\item{anom}{calculate and plot monthly anomalies; for monthly anomalies use anom = "mon", for standardized monthly anomalies use anom ="stmon"}

\item{outname}{specify alternate output filename if input name is not desired}

\item{outtype}{format for output ("png", "pdf")}
}
\description{
This function is for standardized plotitng of indicator time series for Ecosystem Status Reports.
Typical formatting used by Alaska Report Card and CC IEA
Plots time series, values above and below 1 S.D., and mean
Highlights last 5 years of data and shows pattern in mean and trend
}
\examples{
 plotIndicatorTimeSeries("indicator.csv", coltoplot=2:4, plotrownum=3)
 plotIndicatorTimeSeries("amo.csv", coltoplot=2, plotrownum=1, plotcolnum=1, adjlabmain=1, sublabel=F)
 plotIndicatorTimeSeries("aveSocialConnectedness.csv", coltoplot=2, plotrownum=1, plotcolnum=1, sublabel=F)
 plotIndicatorTimeSeries("menhaden_abundance_index.csv", coltoplot=2, plotrownum=1, plotcolnum=1, sublabel=F, yposadj=0.8, widadj=0.8)
 plotIndicatorTimeSeries("seagrass_acreage.csv", coltoplot=2:7, plotrownum=3, plotcolnum=2, adjlabmain=1, sublabel=T, widadj=0.6, trendAnalysis=F)
function()
}
\keyword{}
