% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndicatorTimeSeries.r
\name{plotIndicatorTimeSeries}
\alias{plotIndicatorTimeSeries}
\title{Plotting indicator time series}
\usage{
plotIndicatorTimeSeries(
  indobject,
  coltoplot = 1,
  plotrownum = 1,
  plotcolnum = 1,
  sublabel = F,
  sameYscale = F,
  yposadj = 1,
  widadj = 1,
  hgtadj = 1,
  type = "default",
  CItype = "band",
  trendAnalysis = T,
  tWindow = 5,
  propNAallow = 0.6,
  redgreen = T,
  anom = "none",
  dateformat = "\%b\%Y",
  outname = NA,
  outtype = "",
  ...
)
}
\arguments{
\item{indobject}{an object of class \code{indicatordata}. See details below. For deprecated .csv format, see \link{conv2indicatordata}.}

\item{coltoplot}{an integer or integer list defining the column numbers of indicator file to plot.  Defaults to a single column of data in column 1.}

\item{plotrownum}{an integer defining the number of rows of plots in a multi-panel plot.}

\item{plotcolnum}{an integer defining the number of columns of plots in multi-panel plot.}

\item{sublabel}{a logical value indicating whether optional descriptive information should appear within main label.}

\item{sameYscale}{a logical value indicating whether a consistent y-axis scale is desired across multiple panels.}

\item{yposadj}{a number specifying manual adjustment of position of y-axis label; values > 1 move text further from the axis.}

\item{widadj}{expansion factor to adjust the total width of plot.}

\item{hgtadj}{expansion factor to adjust the total height of plot.}

\item{type}{a character string indicating which type of plot is desired. Defaults to points, with lines for consecutive time steps only.
\code{ptsOnly} and \code{allLines} can be specified for only points or only lines, respectively.}

\item{CItype}{is a character string indictating which type of confidence intervals are desired.
Defaults to shaded bands filling in the area between the upper and lower intervals, or use \code{pts} for an interval plot.}

\item{trendAnalysis}{a logical value indicating whether to highlight the trend in mean and slope over last 5 years; defaults to TRUE unless fewer than 5 years of data.}

\item{tWindow}{an integer defining the number of years over which the recent trend analysis should be calculated; defaults to last 5 years.}

\item{propNAallow}{if fraction denoting the allowable proportion of missing values in last 5 years; when the proportion of NAs exceeds this value, trend analysis will not appear (defaults to 0.5)}

\item{redgreen}{a logical value indicating whether to remove red/green shading of anomalies from plot.}

\item{anom}{a character string indicating whether to convert indicator to monthly anomalies.  One of \code{none}, \code{mon} (monthly anomalies) or \code{stmon} (standardized monthly anomalies) can be used.}

\item{dateformat}{a format as defined in \link{strptime} which is used for monthly time steps only.  Can be full date or month/year combination only.}

\item{outname}{a character string specifying alternate output filename; defaults to using the object name.}

\item{outtype}{a character string specifying format for output, if manual saving is not desired.  Options are \code{bmp}, \code{jpeg}, \code{png}, \code{tiff}, or \code{eps}.}

\item{...}{Arguments to be passed to methods such as specifications for \link{plot} or \link{axis}, particularly \code{cex.axis}, \code{cex.main}, and \code{cex.lab} for sizing labels.}
}
\description{
This function is for standardized plotitng of indicator time series, such as those found in NOAA's Ecosystem Status Reports.
The function imports a data frame of indicator values and dates and plots the time series, denoting mean and values above and below one standard deviation from the mean.
An optional trend analysis highlights changes in the mean and slope of the time series in the last 5 years of data (or other specified window).
}
\note{
A deprecated version of this code allowed for .csv format to be input.  These can now be converted using \link{conv2indicatordata}.

Data must be input as a list object of class \code{indicatordata} which contains at least three attributes: \code{labels}, \code{indicators}, and \code{datelist}.
\itemize{
\item \code{labels} is a \link{data.frame} containing up to 3 rows and the number of columns equal to the number of indicators.
\strong{Row 1} specifies the indicator name, \strong{row 2} specifies the measurement unit, and \strong{row 3} specifies a sublabel.
\item \code{indicators} is a \link{data.frame} containing the indicator data, with each column containing values for a given indicator and time step.
\item \code{datelist} is a \link{vector} containing the time steps at which the indicators were measured, in chronological order.
The length of \verb{datelist`` must be equal to the number of rows in }indicators\verb{. \\item Time can be in year (with century), or monthly time step in a variety of formats (e.g, Jan1986, Jan-86, 1986jan), including or excluding day of month. \\item Optional attributes: }ulim\code{and}llim\verb{are represent, respectively, the upper and lower confidence intervals. They must be in the format of a \link{data.frame}, with equal dimensions to}indicators\verb{.  The first column of }ulim\verb{corresponds to the first column of}indicators`, and so on.
}
}
\examples{
## plot a single indicator
 head(menhaden)
 plotIndicatorTimeSeries(menhaden)

 ## plot a four-panel plot of indicator values reported at different locations
 plotIndicatorTimeSeries(bottomDO, coltoplot = 1:4, sublabel = T, sameYscale = T)

 ## plot an indicator, compare with plot of standardized anomalies
 par(mfrow = c(2, 1))
 plotIndicatorTimeSeries(NPP)
 plotIndicatorTimeSeries(NPP, anom = "stmon")

}
\references{
\url{https://www.aoml.noaa.gov/ocd/ocdweb/ESR_GOMIEA/report/GoM_EcosystemStatusReport2017.pdf}
}
